Moved consumers out into their own project
