+++ CdmsBackend.sln
@@ -33,21 +33,25 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cdms.Types.Gvms.V1", "Cdms.
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cdms.Types.Gvms.Mapping.V1", "Cdms.Types.Gvms.Mapping.V1\Cdms.Types.Gvms.Mapping.V1.csproj", "{6F3F9F42-2DB4-4EC4-BED7-078C8FECA19D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cdms.Business", "Cdms.Business\Cdms.Business.csproj", "{517259EB-B064-409E-98EA-F545D443C854}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cdms.Business", "Cdms.Business\Cdms.Business.csproj", "{517259EB-B064-409E-98EA-F545D443C854}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cdms.BlobService", "Cdms.BlobService\Cdms.BlobService.csproj", "{50EE4A5B-8204-4110-B2A1-0B7233284E46}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cdms.BlobService", "Cdms.BlobService\Cdms.BlobService.csproj", "{50EE4A5B-8204-4110-B2A1-0B7233284E46}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cdms.Azure", "Cdms.Azure\Cdms.Azure.csproj", "{3632CB4F-12D6-4315-B61B-5A487BC0D93C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cdms.Azure", "Cdms.Azure\Cdms.Azure.csproj", "{3632CB4F-12D6-4315-B61B-5A487BC0D93C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cdms.SensitiveData.Tests", "Cdms.SensitiveData.Tests\Cdms.SensitiveData.Tests.csproj", "{D74D5739-3BF5-4AFA-AF9E-5C4171A6DE83}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cdms.SensitiveData.Tests", "Cdms.SensitiveData.Tests\Cdms.SensitiveData.Tests.csproj", "{D74D5739-3BF5-4AFA-AF9E-5C4171A6DE83}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cdms.Business.Tests", "Cdms.Business.Tests\Cdms.Business.Tests.csproj", "{6ECCF74E-C518-461F-8E00-9E1CD38304C2}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cdms.Business.Tests", "Cdms.Business.Tests\Cdms.Business.Tests.csproj", "{6ECCF74E-C518-461F-8E00-9E1CD38304C2}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestDataGenerator", "TestDataGenerator\TestDataGenerator.csproj", "{BBCE956A-B91B-487D-9AAF-0FDC69FD193B}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestDataGenerator.Tests", "TestDataGenerator.Tests\TestDataGenerator.Tests.csproj", "{DE402A21-9EC4-4E00-838C-DD76C3F50DE2}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "types", "types", "{3F5B86A8-8142-40C3-9525-5AA00DEDEF34}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cdms.Consumers", "Cdms.Consumers\Cdms.Consumers.csproj", "{2BAE849A-6CCA-481E-8094-1A59B32D86A4}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cdms.Consumers.Tests", "Cdms.Consumers.Tests\Cdms.Consumers.Tests.csproj", "{91F2D8AC-0AF3-424C-805C-CEC0F8EB0BF7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -135,6 +139,14 @@ Global
 		{DE402A21-9EC4-4E00-838C-DD76C3F50DE2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DE402A21-9EC4-4E00-838C-DD76C3F50DE2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DE402A21-9EC4-4E00-838C-DD76C3F50DE2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2BAE849A-6CCA-481E-8094-1A59B32D86A4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2BAE849A-6CCA-481E-8094-1A59B32D86A4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2BAE849A-6CCA-481E-8094-1A59B32D86A4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2BAE849A-6CCA-481E-8094-1A59B32D86A4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{91F2D8AC-0AF3-424C-805C-CEC0F8EB0BF7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{91F2D8AC-0AF3-424C-805C-CEC0F8EB0BF7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{91F2D8AC-0AF3-424C-805C-CEC0F8EB0BF7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{91F2D8AC-0AF3-424C-805C-CEC0F8EB0BF7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
